Begin better framework for syntax highlighting.
Mostly rely on pegs, but used a better data structure
for coloring spans. Highlighters no should be pegs that
take a string of text and return an array of spans. A span
is either a string for un-highlighted text, or a pair
(class text) for highlighted text. The class is a keyword
label that is mapped to a color during rendering.